clean code and leave hexagonal exaple at the moment
