+++ src/internals/adapters/handlers/routing.go
@@ -5,8 +5,8 @@ import (
 	"io"
 	"net/http"
 
-	"github.com/gcarrenho/guidemysteps/src/internal/core/models"
-	"github.com/gcarrenho/guidemysteps/src/internal/core/ports"
+	"github.com/gcarrenho/guidemysteps/src/internals/core/models"
+	"github.com/gcarrenho/guidemysteps/src/internals/core/ports"
 
 	"github.com/gin-gonic/gin"
 )
