+++ src/internals/adapters/repositories/open_street_map/models/osm_routes.go
