+++ src/internals/adapters/repositories/open_street_map/osm.go
@@ -8,8 +8,8 @@ import (
 	"net/http"
 	"time"
 
-	osmmodels "github.com/gcarrenho/guidemysteps/src/internal/adapters/repositories/open_street_map/models"
-	"github.com/gcarrenho/guidemysteps/src/internal/core/models"
+	osmmodels "github.com/gcarrenho/guidemysteps/src/internals/adapters/repositories/open_street_map/models"
+	"github.com/gcarrenho/guidemysteps/src/internals/core/models"
 )
 
 type routingRepo struct {
