+++ src/internals/core/services/routing.go
@@ -1,9 +1,9 @@
 package services
 
 import (
-	repositories "github.com/gcarrenho/guidemysteps/src/internal/adapters/repositories/open_street_map/models"
-	"github.com/gcarrenho/guidemysteps/src/internal/core/models"
-	"github.com/gcarrenho/guidemysteps/src/internal/core/ports"
+	repositories "github.com/gcarrenho/guidemysteps/src/internals/adapters/repositories/open_street_map/models"
+	"github.com/gcarrenho/guidemysteps/src/internals/core/models"
+	"github.com/gcarrenho/guidemysteps/src/internals/core/ports"
 )
 
 const (
